Base source for Pribambase WIN64
STARTING INFO: The base source code and extensions for pribambase on the windows64 platform. Since this is a new drop for this, I am unaware of how its set up on multiple platforms. I am sure there are macros that directly conduct the scripts to run for different operating environments. Since this is a community effort, I'll try to comment as much as I can in order to bring clarity to the code presented so its easier for others viewing and editing. 
PLEASE NOTE: This is the base version of this script from the original creator. They have informed me it is extremely messy and "fragile", please keep rude notes to yourself. We're trying to make this better, it is already amazing that we have something like this to aid low poly work in the first place. Thanks!