Fix sync on UV with blender 3.x
Fixes a sync issue with blender 3.4 and newer specifically. Tested working on 3.0 will updated on 4.0 and 4.x status. 
This fixes a sync issue with the UV send routine causing instability or send errors. UV maps should now update appropriately when working on the blender side. This solution works on all platforms.
Thanks ~Dmitrii for the fix, your contribution is much appreciated!